// src/api/models/TestResultStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum TestResultStatus {
    SUCCESS = 'SUCCESS',
    SKIPPED = 'SKIPPED',
    PARTIAL_SKIPPED = 'PARTIAL_SKIPPED',
    FAILED = 'FAILED',
    SYSTEM_ERROR = 'SYSTEM_ERROR',
    NO_CORRESPONDING_TEST = 'NO_CORRESPONDING_TEST',
    NO_REPORT_ERROR = 'NO_REPORT_ERROR',
    UNKNOWN = 'UNKNOWN',
    CANCELED = 'CANCELED',
}